+++ .env
@@ -1,2 +1,2 @@
-MONGODB_URI="mongodb+srv://marv:5020@mo-db.dsbas.mongodb.net/petsInn?retryWrites=true"
+MONGODB_URI="mongodb+srv://tedNorthcoders:PetInn0369@petinn.gyiq5ap.mongodb.net/PetInn?retryWrites=true&w=majority"
 SECRET="pets"
