+++ app.js
@@ -3,9 +3,10 @@ const app = express();
 const cors = require("cors");
 const dotenv = require("dotenv");
 const session = require("express-session");
-const passport = require("./passport");
+const passport = require("./middleware/passport");
 const bcryptJs = require("bcryptjs");
-const User = require("./user.model");
+const userRouters = require("./routes/user.routes");
+const apiRouter = require("./routes/api.routes");
 
 dotenv.config();
 app.use(express.json());
@@ -27,30 +28,7 @@ app.use((req, res, next) => {
 });
 
 // Routes
-app.post("/signup", (req, res) => {
-  bcryptJs.hash(req.body.password, 10, (err, hashedPassword) => {
-    new User({
-      username: req.body.username,
-      password: hashedPassword,
-    }).save((err) => {
-      if (err) console.log(err);
-    });
-  });
-
-  res.send({ user: req.body });
-});
-
-app.post("/login", passport.authenticate("local"), (req, res) => {
-  res.send({ username: req.user.username });
-});
-
-app.get("/logout", (req, res) => {
-  req.logout((err) => {
-    if (err) {
-      return next(err);
-    }
-    res.sendStatus(200);
-  });
-});
+app.use("/api", apiRouter);
+app.use("/api/users", userRouters);
 
 module.exports = app;
