+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon listen.js"
+    "start": "nodemon -L listen.js"
   },
   "keywords": [],
   "author": "",
