+++ routes/api.routes.js
@@ -1,0 +1,6 @@
+const express = require("express");
+const { model } = require("mongoose");
+
+const router = express.Router();
+
+module.exports = router;
